Brand and Customer added

+++ src/main/java/com/aspirecoders/instock/repository/BrandRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aspirecoders.instock.model.Brand;
 
-public interface BrandRepo extends JpaRepository<Brand, Integer> {
-
+public interface BrandRepo extends JpaRepository<Brand, Long> {
+  // You can add custom query methods here if needed
 }
